# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_alemancenter_fiber-api_internal_models.Event:
    properties:
      created_at:
        type: string
      description:
        type: string
      event_date:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.DatabaseInfo:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.EventInput:
    properties:
      description:
        type: string
      event_date:
        type: string
      title:
        maxLength: 500
        minLength: 2
        type: string
    required:
    - event_date
    - title
    type: object
...
  github_com_alemancenter_fiber-api_internal_utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      pagination:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      success:
        type: boolean
      token:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_utils.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
host: api.alemancenter.com
info:
  contact:
    email: support@alemancenter.com
    name: API Support
  description: Backend API for Alemancenter Educational Platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Alemancenter API
  version: 2.0.0
paths:
...
  /dashboard/calendar/databases:
    get:
      description: Returns available calendar databases/countries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.DatabaseInfo'
                  type: array
              type: object
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: List Calendar Databases
      tags:
      - Calendar
  /dashboard/calendar/events:
    get:
      description: Returns a list of calendar events for dashboard management, optionally
        filtered by date range
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Event'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: List Events (Dashboard)
      tags:
      - Calendar
    post:
      consumes:
      - application/json
      description: Create a new calendar event
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: Event data payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.EventInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Create Event
      tags:
      - Calendar
  /dashboard/calendar/events/{id}:
    delete:
      description: Delete an existing calendar event by ID
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Delete Event
      tags:
      - Calendar
    put:
      consumes:
      - application/json
      description: Update an existing calendar event by ID
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.EventInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      - FrontendKeyAuth: []
      summary: Update Event
      tags:
      - Calendar
...
  /home/calendar:
    get:
      description: Returns upcoming public calendar events
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Event'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      summary: List Public Events
      tags:
      - Calendar
  /home/event/{id}:
    get:
      description: Returns details of a specific public calendar event by ID
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      summary: Get Public Event
      tags:
      - Calendar
...
schemes:
- https
- http
security:
- FrontendKeyAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  FrontendKeyAuth:
    description: Frontend identifier key for public endpoints.
    in: header
    name: X-Frontend-Key
    type: apiKey
swagger: "2.0"